requirements.txt
aiogram~=2.19
pydantic~=1.9.0
python-dateutil~=2.8.2
git+https://github.com/Egorrko/aioping